backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
gunicorn==21.2.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.28.0
alembic==1.12.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic-settings==2.1.0
redis==5.0.1
email-validator==2.1.0.post1
httpx==0.25.2
pytest==7.4.4
pytest-asyncio==0.23.2
black==23.12.1
ruff==0.1.9
mypy==1.7.1
